fix: energy accumulation logic, precision, and initialization
Refactored the energy calculation architecture in _updateTimes() to use incremental accumulation and resolved several related bugs that caused data inflation and precision loss.

Specific changes include:

- Fix Total Energy Calculation: Prevented _energy_total_kWh from being overwritten and recalculated from scratch every loop. It now properly accumulates incrementally alongside the daily counter.

- Guard Daily Accumulation: Moved the daily energy accumulation (_energy_daily_Ws) inside the NTP time guard so it no longer blindly accumulates before the system clock syncs.

- Prevent Millisecond Truncation: Replaced = 0 with %= 1000 when flushing sub-second timers in _updateTimes() and saveSettings(). This preserves the remainder and prevents the silent loss of ~48 seconds per day.

- Improve Long-term Precision: Upgraded _energy_total_kWh from float to double to prevent significant-digit degradation after months of continuous use.

- Initialize Class Members: Added explicit initialization for all energy-related variables and millisecond counters in the BWC constructor to prevent undefined behavior.

- Add Safe JSON Defaults: Added | 0.0 fallbacks in _loadSettings() to safely handle missing keys from older firmware configurations.

+++ Code/lib/BWC_unified/bwc.h
@@ -197,9 +197,9 @@ private:
     uint32_t _wc_interval;
     uint32_t _ph_interval;         // pH check interval in days
     bool _audio_enabled;
-    float _energy_total_kWh;
-    double _energy_daily_Ws; // Wattseconds internally
-    int _energy_daily_yday = -1; // day-of-year (0-365) when daily counter was last valid; -1 = unknown
+    double _energy_total_kWh;
+    double _energy_daily_Ws;
+    int _energy_daily_yday = -1;
     int _energy_power_W;
     String _plz;
     bool _weather = 0;
